requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.10.4
pydantic-settings==2.6.1
celery==5.4.0
redis==5.0.6
structlog==24.1.0
python-multipart==0.0.9
# Bump to versions compatible with Python 3.13
sqlmodel==0.0.21
SQLAlchemy==2.0.36
psycopg[binary]==3.2.9
alembic==1.13.2
pytest==8.3.3
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
